moved route config to route.rs file

+++ server/src/model/repository.rs
@@ -29,7 +29,7 @@ pub struct Repository {
   #[serde(rename = "_id")]
   pub _id: bson::oid::ObjectId,
   pub description: Option<String>,
-  pub fork_count: u32,
+  pub fork_count: f64,
   #[serde(skip_serializing_if = "Option::is_none")]
   pub license_info: Option<License>,
   pub name: String,
